+++ src/main/webapp/view/css/article/article.css
@@ -1,0 +1,55 @@
+@CHARSET "UTF-8";
+
+.panel-heading .btn, .list-group .btn {
+	float: right;
+	margin-right: 5px;
+	padding: 0;
+}
+
+.callout-danger {
+	margin-bottom: 15px;
+	display: inline-block;
+	padding: 2px 12px;
+	border-left: 3px solid #ce4844;
+}
+
+.article-panel .panel-heading {
+	padding-right: 125px;
+	position: relative;
+}
+
+.article-panel .panel-heading .pull-right {
+	position: absolute;
+	right: 20px;
+	top: 10px;
+}
+
+article {
+	padding: 0 35px 40px;
+}
+
+category {
+	display: block;
+	padding: 10px 20px;
+	margin-bottom: 15px;
+	background-color: rgba(128, 128, 128, 0.1);
+}
+
+summary {
+	padding: 0 35px 20px;
+}
+
+summary pre {
+	border-left: 10px solid rgba(128, 128, 128, 0.08);
+}
+
+._loading {
+	display: none;
+	position: absolute;
+	top: 0;
+	width: 100%;
+	height: 100%;
+	z-index: 999;
+	background-color: #88a825;
+	opacity: 0.6;
+}
