+++ src/main/webapp/view/js/article/articleList.js
@@ -20,125 +20,11 @@
  */
 require([
     "jquery",
-    'underscore',
-    'nprogress',
-    "jBoxUtil",
-    "commonutils",
+    "js/article/providers/articleListProvider",
     "js/article/articleWidget",
-    "js/article/article",
-    "bootstrap",
-    'pagination'
-], function ($, _, NProgress, jBoxUtil, commonutils, articleWidget) {
-    var pageSize = 2;
-    var currentPage = 0;
-    var totalPages = 0;
-
-    var articleListHTML = '<article class="list-group-item" data-uid="<%= articleUid %>">' +
-        '<h4><a class="article-pjax-href text-info" href="/article/<%= articleUid %>"><%= articleTitle %></a></h4>' +
-        '<section title="<%= articleSummary %>"><%= articleSummary %></section>' +
-        '<div>' +
-        '<small>' +
-        '<%= articleHits %>人阅读&nbsp;' +
-        '<i class="icon-calendar"></i><%= articleCreateTime %>' +
-        '</small>' +
-        '<button data-uid="<%= articleUid %>" type="button" class="btn btn-danger btn-xs btn-list-article-recycle"><i class="glyphicon glyphicon-trash"></i>删除</button>' +
-        '<button data-uid="<%= articleUid %>" type="button" class="btn btn-warning btn-xs btn-list-article-withdraw"><i class="glyphicon glyphicon-share"></i>撤回</button>' +
-        '<a href="/editor/<%= articleUid %>" type="button" class="btn btn-primary btn-xs btn-article-edit"><i class="glyphicon glyphicon-edit"></i>编辑</a>' +
-        '</div>' +
-        '</article>';
-
-    var articleListAnonymousHTML = '<article class="list-group-item" data-uid="<%= articleUid %>">' +
-        '<h4><a class="article-pjax-href text-info" href="/article/<%= articleUid %>"><%= articleTitle %></a></h4>' +
-        '<section title="<%= articleSummary %>"><%= articleSummary %></section>' +
-        '<div>' +
-        '<small>' +
-        '<%= articleHits %>人阅读&nbsp;' +
-        '<i class="icon-calendar"><%= articleCreateTime %></i>' +
-        '</small>' +
-        '</div>' +
-        '</article>';
-
-    $('.pagination.article-pagination').jqPagination({
-        page_string: '第{current_page}页, 共{max_page}页',
-        current_page: 1,
-        max_page: null,
-        paged: function (page) {
-            if (page == currentPage + 1) {
-                return;
-            }
-
-            searchFunc(page - 1);
-        }
-    });
-
-    var searchFuncs = {};
-    searchFuncs.article = function (page) {
-        searchArticles("/article/query/", page);
-    };
-    searchFuncs.category = function (page) {
-        searchArticles("/article/category/query/", page);
-    };
-    searchFuncs.archive = function (page) {
-        searchArticles("/article/archive/query/", page);
-    };
-    searchFuncs.search = function (page) {
-
-    };
-
-    var funcname = $("#funcname").val();
-    var funcparam = $("#funcparam").val();
-
-    var func = searchFuncs[funcname];
-    var searchFunc = typeof func == "function" ? func : function () {
-        console.warn("Cannot Find Any Search Function!");
-    };
-
-    function searchArticles(url, page) {
-        NProgress.start();
-        $.ajax({
-            url: url + funcparam,
-            async: true,
-            type: 'post',
-            cache: false,
-            dataType: 'json',
-            data: {state: 'PUBLISHED', page: page, size: pageSize},
-            success: function (data, textStatus, XMLHttpRequest) {
-                if (null != data.errmsg) {
-                    // 出现错误
-                    jBoxUtil.noticeError({content: data.errmsg});
-                    return;
-                }
-
-                reShowArticleList(data.articles, page, data.totalPages);
-            },
-            error: function () {
-                jBoxUtil.noticeError({content: "未知错误"});
-            },
-            complete: function () {
-                NProgress.done();
-            }
-        });
-    }
-
-    function reShowArticleList(_articles, _page, _totalPages) {
-        var list = [];
-        _.each(_articles, function (article) {
-            list.push(_.template(!!logined ? articleListHTML : articleListAnonymousHTML)({
-                articleCreateTime: commonutils.dateFormate(new Date(article.createTime)),
-                articleTitle: article.title,
-                articleSummary: article.summary,
-                articleHits: article.hits,
-                articleUid: article.uid
-            }));
-        });
-
-        $(".article-list").empty().append(list.join(""));
-
-        currentPage = _page;
-        totalPages = _totalPages < 1 ? 1 : _totalPages;
-        /*$('.pagination.article-pagination').jqPagination("option", "current_page", currentPage + 1);*/
-        $('.pagination.article-pagination').jqPagination("option", "max_page", totalPages);
-    }
-
-    searchFunc(0);
+    'js/article/providers/articlePjaxProvider',
+    "bootstrap"
+], function ($, articleList, articleWidget) {
+    articleList.initArticleList();
+    window.setTimeout(articleWidget.init, 1000);
 });
