+++ src/main/webapp/view/js/article/articleWidget.js
@@ -31,9 +31,9 @@ define([
     /*********************
      * 文章分类
      *********************/
-    var categoryListHTML = '<a href="/article/category/<%= categoryName %>" class="has-badge"><span class="text-overflow"><i class="glyphicon glyphicon-tags"></i> <%= categoryName %></span><span class="badge"><%= articleNum %></span></a>';
-    var archiveListHTML = '<a href="/article/archive/<%= archiveDate %>" class="has-badge"><span class="text-overflow"><i class="glyphicon glyphicon-calendar"></i> <%= archiveDate %></span><span class="badge"><%= articleNum %></span></a>';
-    var hitsListHTML = '<a href="/article/<%= articleUid %>" class="article-pjax-href has-badge" title="[<%= articleTitle %>]\n<%= articleSummary %>"><span class="text-overflow"><i class="glyphicon glyphicon-calendar"></i> <%= articleTitle %></span><span class="badge"><%= articleHits %></span></a>';
+    var categoryListHTML = '<a href="/article/category/<%= categoryName %>" class="article-list-pjax-href has-badge"><span class="text-overflow"><i class="glyphicon glyphicon-tags"></i> <%= categoryName %></span><span class="badge"><%= articleNum %></span></a>';
+    var archiveListHTML = '<a href="/article/archive/<%= archiveDate %>" class="article-list-pjax-href has-badge"><span class="text-overflow"><i class="glyphicon glyphicon-calendar"></i> <%= archiveDate %></span><span class="badge"><%= articleNum %></span></a>';
+    var hitsListHTML = '<a href="/article/<%= articleUid %>" class="article-content-pjax-href has-badge" title="[<%= articleTitle %>]\n<%= articleSummary %>"><span class="text-overflow"><i class="glyphicon glyphicon-calendar"></i> <%= articleTitle %></span><span class="badge"><%= articleHits %></span></a>';
 
     function initCategory() {
         $.ajax({
