+++ src/main/webapp/view/pages/article/article.jsp
@@ -10,20 +10,8 @@
 <html>
 <head>
 	<link rel="stylesheet" href="<c:url value='/view/style/themes/default.css?v=${version}'/>" type="text/css">
-
+	<link rel="stylesheet" href="<c:url value='/view/css/article/article.css?v=${version}'/>" type="text/css">
     <title><c:out value="${fn:substring(title,0,32)}" />·MBLOG</title>
-    
-    <style>
-    	.panel-heading .btn, .list-group .btn {float: right; margin-right: 5px; padding: 0;}
-    	.callout-danger {margin-bottom: 15px;display: inline-block;padding: 2px 12px;border-left: 3px solid #ce4844;}
-    	.article-panel .panel-heading {padding-right: 125px; position: relative;}
-    	.article-panel .panel-heading .pull-right {position: absolute;right: 20px;top: 10px;}
-    	article {padding: 0 35px 40px;}
-    	category {display: block;padding: 10px 20px; margin-bottom: 15px;background-color: rgba(128, 128, 128, 0.1);}
-    	summary {padding: 0 35px 20px;}
-    	summary pre {border-left: 10px solid rgba(128, 128, 128, 0.08);}
-    	._loading{display:none;position:absolute;top:0;width:100%;height:100%;z-index:999;background-color:#88a825;opacity:0.6;}
-    </style>
 </head>
 <body>
     <c:set var="position" value="relative" scope="request"></c:set>
@@ -44,6 +32,13 @@
     </div>
     
 	<c:import url="../nav/footer.jsp" />
+	
+	<script>
+		var logined = false;
+		<sec:authorize access="hasAnyRole('ROLE_ADMIN,ROLE_USER')">
+			logined = true;
+		</sec:authorize>
+	</script>
 	
 	<script type="text/javascript" src="https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS_HTML"></script>
 	<script src="<c:url value="/view/plugins/requirejs/require.js?v=${version}" />"></script>
