+++ src/main/webapp/view/pages/article/searchListPjax.jsp
@@ -1,0 +1,13 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
+
+<%@ taglib prefix="c" uri="c_tld" %><!-- http://java.sun.com/jsp/jstl/core -->
+<%@ taglib prefix="fmt" uri="fmt_tld" %><!-- http://java.sun.com/jsp/jstl/functions -->
+<%@ taglib prefix="fn" uri="fn_tld" %><!-- http://java.sun.com/jsp/jstl/fmt -->
+<%@ taglib prefix="spring" uri="spring_tld" %> <!-- http://www.springframework.org/tags -->
+<%@ taglib prefix="sec" uri="security_tld" %> <!-- http://www.springframework.org/security/tags" -->
+
+<c:if test="${pjax}" >
+	<title><c:out value="${keywords}"/>·MBLOG</title>
+</c:if>
+
+<div class="list-group article-list"><h1 style='color: lightgray;'>请稍后</h1></div>
