+++ src/main/webapp/view/style/styles/basic.less
@@ -561,3 +561,13 @@ article.list-group-item {
 .quick-access {
   margin: 20px 0 0 160px;
 }
+
+/********************
+ * goup
+ ********************/
+.goup-container {
+  opacity: 0.5;
+  &:hover {
+    opacity: 1;
+  }
+}
